Split euphonic.cli.utils to submodules
This file is way too large and hard to navigate
